add public k8s helper functions
