+++ sdk/python/kfp_tekton/compiler/_k8s_helper.py
@@ -100,7 +100,7 @@ def convert_k8s_obj_to_json(k8s_obj):
       if isinstance(k8s_obj.value, str):
         return k8s_obj.value
       return '$(inputs.params.%s)' % k8s_obj.full_name  # change for Tekton
-    
+
     if isinstance(k8s_obj, dict):
       obj_dict = k8s_obj
     else:
@@ -187,7 +187,7 @@ def sanitize_k8s_object(k8s_obj, type=None):
                     for sub_obj in k8s_obj)
     elif isinstance(k8s_obj, (datetime, date)):
       return k8s_obj
-    
+
     if isinstance(k8s_obj, dict):
       return k8s_obj
     else:
